fix compile + remove dsp_process and move content to synth_engine + updated docs for details
