+++ logue-sdk-ripplerx.code-workspace
@@ -1,0 +1,19 @@
+{
+    "launch": {
+        "configurations": [
+            {
+                "type": "cmake",
+                "request": "launch",
+                "name": "Debug portfile(s)",
+                "cmakeDebugType": "external",
+                "pipeName": "\\\\.\\pipe\\vcpkg_ext_portfile_dbg",
+                "preLaunchTask": "Debug vcpkg commands"
+            }
+        ]
+    },
+    "folders": [
+        {
+            "path": "."
+        }
+    ]
+}
