+++ platform/drumlogue/ripplerx/PROGRESS.md
@@ -7,28 +7,39 @@
 ## Phase 2: API & UI Binding [COMPLETED]
 - [x] Connect `header.c` parameter indices to the engine.
 - [x] Write `setParameter` translation logic (Pre-calculate all filter coefficients in the UI thread).
-- [x] Implement the dynamic Sample Loading bounds check to prevent RTOS crashes.
 
 ## Phase 3: The Audio Processing Loop [COMPLETED]
 - [x] Write the per-sample DSP loop (Fractional Delay -> Filter -> Feedback -> Write).
 - [x] Implement linear interpolation for accurate pitch tuning.
 - [x] Write the master block processing loop (Summation & Exciter routing).
 
-## Phase 3: The Audio Processing Loop [COMPLETED]
-## Phase 4: The Hardware "Hello World" [READY FOR TEST]
-- [x] `unit.cc` wrapper implemented with complete OS lifecycle compliance (`Init`, `Reset`, etc.).
-- [ ] Compile the minimal viable engine.
-- [ ] Deploy to Drumlogue and verify CPU load and stability.
+## Phase 4: Architectural Refactor & Sample Management [COMPLETED]
 
-## Phase 5: Envelopes & Exciters [READY FOR TEST]
-- [x] `envelope.h`: Implement a fast, branchless AR/ADSR struct.
-- [x] `noise.h`: Add a fast PRNG (Xorshift32) noise generator.
-- [x] **Incremental Integration:** Modules added to `dsp_core.h` and `dsp_process.h` behind `#ifdef ENABLE_PHASE_5_EXCITERS` for safe debugging.
+### Summary
+The initial implementation of sample loading was found to be unreliable, leading to silence. The architecture has been refactored based on the working 'resonator' reference project to improve stability and robustness.
+
+### Changes & Fixes
+- **[FIXED]** Replaced the fragile, state-based sample loading mechanism with a "just-in-time" model. The `NoteOn` function now directly fetches sample data from the OS when a note is triggered.
+- **[REMOVED]** Deleted the `loadConfigureSample` and `clearSampleState` functions, which were triggered by the UI, in favor of the more robust `NoteOn`-based loading.
+- **[COMPLETED]** Added defensive checks to the `Init` and `NoteOn` functions to ensure all OS-provided sample API pointers are valid before use, preventing potential crashes.
+
+## Phase 5: Sound Generation & UI Integration [IN PROGRESS]
+
+### Next Steps (To-Do)
+- **[ ] Synthesizer First:** The synth engine should produce sound without samples, using the "Mallet" and "Noise" exciters.
+    - **[ ] Enable Synthesis:** Uncomment the `#define ENABLE_PHASE_5_EXCITERS` in `dsp_core.h` to activate the existing noise generator.
+    - **[ ] Implement Mallet:** Implement the mathematical "Mallet" exciter described in `README.md`.
+    - **[ ] Implement Mixer:** Add a parameter to select or mix between Sample, Noise, and Mallet exciters.
+
+- **[ ] Full UI Integration:**
+    - **[ ] Implement `unit_get_param_value`:** In `unit.cc`, fully implement this function to report the current state of all parameters back to the UI.
+    - **[ ] Implement `unit_get_param_str_value`:** In `unit.cc`, implement this to display the correct names for string-based parameters like the current sample, bank, and exciter type.
+    - **[ ] Test & Verify:** Thoroughly test that all UI controls map correctly to the engine and that the screen accurately reflects the state of the synth.
 
 ## Phase 6: Filters & Master FX [IN PROGRESS]
 - [x] `filter.h`: Implement a 2-pole Chamberlin State Variable Filter (SVF).
-- [x] Tie `header.c` Cutoff and Resonance to the Master SVF.
-- [x] Integrate SVF incrementally into the audio loop.
+- [ ] Tie `header.c` Cutoff and Resonance to the Master SVF.
+- [ ] Integrate SVF incrementally into the audio loop.
 
 ## Phase 7: Waveguide Models & Tables [PENDING]
 - [ ] `constants.h` & `tables.h`: Define fast-math lookup tables and tuning constants.
@@ -37,5 +48,3 @@
 ## Phase 8: Preset Design & Acoustic Tuning [PENDING]
 - [ ] Reverse-engineer waveguide parameters for original 28 presets.
 - [ ] Derive exact physics coefficients for new instruments: Timpani, Djambé, Taiko Drum, Marching Snare, Tam Tam, Koto.
-
-
