+++ platform/drumlogue/ripplerx/project.mk
@@ -1,0 +1,6 @@
+PROJECT_NAME := ripplerx
+PROJECT_TYPE := synth
+PROJECT_DEPS := 
+PROJECT_CATEGORY := 
+PROJECT_FOLDER := drumlogue/$(PROJECT_NAME)
+include ../../common/common.mk
